rewrite ClassData::get_instance_method()
Previously, we used this approach:

* Collect all ancestors into `std::vector` (interfaces, base classes)
* Iterate over them, return the first member match

What we get wrong with this approach is that we could get an abstract method
instead of the implementation as the order of `get_all_ancestors()` doesn't
prefer classes to interfaces.

Consider this case:

```php
interface Iface { function f(); }

// Actually implements Iface, but doesn't state that
abstract class AbstractBase {
  public function f() {}
}

abstract class AbstractBase2 extends AbstractBase {}

// Implements Iface via AbstractBase, which is a base class of AbstractBase2
class Concrete extends AbstractBase2 implements Iface {}
```

With the old approach, we'll get a compile-time error.

The new approach prefers non-abstract methods; it'll traverse the classes and
try to find the first non-abstract method. If it's not possible, it will
return the first abstract method (like we did before). The code is slightly
more complicated than that just to avoid redundant work.

As we do more (useful) work now and sometimes need to check more classes,
I tried to write this method so it's not slower than before.

	old method:
	| Instance method lookup | 365669 | 3.478 sec | 6.370 sec | - | +6.3 MB |
	new method:
	| Instance method lookup | 365669 | 3.256 sec | 7.154 sec | - |       - |

Note: locking is applied in the same fashion as in find_by_local_name, per `.members` lookup.

Fixes #215